server/requirements.txt
# For the Web Server
fastapi
uvicorn
gunicorn
python-multipart

# For Machine Learning & Image Processing
face_recognition
Pillow
numpy

# For Database and Authentication 
pymongo[srv]
bcrypt
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings

# Google Cloud Storage
google-cloud-storage